requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0

numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
pdfplumber>=0.11.0
openpyxl>=3.1.0

fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0

python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.12.0
